Moved process selection type to own file
