+++ src/processTargets/index.ts
@@ -1,23 +1,18 @@
-import { concat, zip } from "lodash";
-import { Position, Range, Selection, TextDocument } from "vscode";
+import { zip } from "lodash";
+import { Selection } from "vscode";
 import { performInsideOutsideAdjustment } from "../performInsideOutsideAdjustment";
-import {
-  selectionFromPositions,
-  selectionWithEditorFromPositions,
-} from "../selectionUtils";
 import {
   LineNumberPosition,
   Mark,
-  Modifier,
   PrimitiveTarget,
   ProcessedTargetsContext,
   RangeTarget,
-  SelectionContext,
   SelectionWithEditor,
   Target,
   TypedSelection,
 } from "../Types";
 import processModifier from "./processModifier";
+import processSelectionType from "./processSelectionType";
 
 export default function (
   context: ProcessedTargetsContext,
@@ -179,16 +174,12 @@ function processPrimitiveTarget(
   target: PrimitiveTarget
 ): TypedSelection[] {
   const markSelections = getSelectionsFromMark(context, target.mark);
-  const transformedSelections = concat(
-    [],
-    ...markSelections.map((markSelection) =>
-      processModifier(context, target, markSelection)
-    )
+  const modifiedSelections = markSelections.flatMap((markSelection) =>
+    processModifier(context, target, markSelection)
   );
-
-  const typedSelections = transformedSelections.map(
+  const typedSelections = modifiedSelections.map(
     ({ selection, context: selectionContext }) =>
-      createTypedSelection(context, target, selection, selectionContext)
+      processSelectionType(context, target, selection, selectionContext)
   );
   return typedSelections.map((selection) =>
     performPositionAdjustment(context, target, selection)
@@ -264,122 +255,6 @@ function getSelectionsFromMark(
   }
 }
 
-function createTypedSelection(
-  context: ProcessedTargetsContext,
-  target: PrimitiveTarget,
-  selection: SelectionWithEditor,
-  selectionContext: SelectionContext
-): TypedSelection {
-  const { selectionType, insideOutsideType, position, modifier } = target;
-  const { document } = selection.editor;
-
-  switch (selectionType) {
-    case "token":
-      return {
-        selection,
-        selectionType,
-        position,
-        insideOutsideType,
-        selectionContext: getTokenSelectionContext(
-          selection,
-          modifier,
-          selectionContext
-        ),
-      };
-
-    case "line": {
-      const startLine = document.lineAt(selection.selection.start);
-      const endLine = document.lineAt(selection.selection.end);
-      const start = new Position(
-        startLine.lineNumber,
-        startLine.firstNonWhitespaceCharacterIndex
-      );
-      const end = endLine.range.end;
-
-      const newSelection = selectionWithEditorFromPositions(
-        selection,
-        start,
-        end
-      );
-
-      return {
-        selection: newSelection,
-        selectionType,
-        position,
-        insideOutsideType,
-        selectionContext: getLineSelectionContext(newSelection),
-      };
-    }
-
-    case "document": {
-      const firstLine = document.lineAt(0);
-      const lastLine = document.lineAt(document.lineCount - 1);
-      const start = firstLine.range.start;
-      const end = lastLine.range.end;
-      const newSelection = selectionWithEditorFromPositions(
-        selection,
-        start,
-        end
-      );
-
-      return {
-        selection: newSelection,
-        selectionType,
-        position,
-        insideOutsideType,
-        selectionContext,
-      };
-    }
-
-    case "paragraph": {
-      let startLine = document.lineAt(selection.selection.start);
-      if (!startLine.isEmptyOrWhitespace) {
-        while (startLine.lineNumber > 0) {
-          const line = document.lineAt(startLine.lineNumber - 1);
-          if (line.isEmptyOrWhitespace) {
-            break;
-          }
-          startLine = line;
-        }
-      }
-      const lineCount = document.lineCount;
-      let endLine = document.lineAt(selection.selection.end);
-      if (!endLine.isEmptyOrWhitespace) {
-        while (endLine.lineNumber + 1 < lineCount) {
-          const line = document.lineAt(endLine.lineNumber + 1);
-          if (line.isEmptyOrWhitespace) {
-            break;
-          }
-          endLine = line;
-        }
-      }
-
-      const start = new Position(
-        startLine.lineNumber,
-        startLine.firstNonWhitespaceCharacterIndex
-      );
-      const end = endLine.range.end;
-
-      const newSelection = selectionWithEditorFromPositions(
-        selection,
-        start,
-        end
-      );
-
-      return {
-        selection: newSelection,
-        position,
-        selectionType,
-        insideOutsideType,
-        selectionContext: getParagraphSelectionContext(newSelection),
-      };
-    }
-
-    case "character":
-      throw new Error("Not implemented");
-  }
-}
-
 function performPositionAdjustment(
   context: ProcessedTargetsContext,
   target: PrimitiveTarget,
@@ -417,181 +292,4 @@ function performPositionAdjustment(
     insideOutsideType: target.insideOutsideType ?? null,
     position,
   };
-}
-
-function getTokenSelectionContext(
-  selection: SelectionWithEditor,
-  modifier: Modifier,
-  selectionContext: SelectionContext
-): SelectionContext {
-  if (!isSelectionContextEmpty(selectionContext)) {
-    return selectionContext;
-  }
-  if (modifier.type === "subpiece") {
-    return selectionContext;
-  }
-
-  const document = selection.editor.document;
-  const { start, end } = selection.selection;
-
-  const startLine = document.lineAt(start);
-  const leadingText = startLine.text.slice(0, start.character);
-  const leadingDelimiters = leadingText.match(/\s+$/);
-  const leadingDelimiterRange =
-    leadingDelimiters != null
-      ? new Range(
-          start.line,
-          start.character - leadingDelimiters[0].length,
-          start.line,
-          start.character
-        )
-      : null;
-
-  const endLine = document.lineAt(end);
-  const trailingText = endLine.text.slice(end.character);
-  const trailingDelimiters = trailingText.match(/^\s+/);
-  const trailingDelimiterRange =
-    trailingDelimiters != null
-      ? new Range(
-          end.line,
-          end.character,
-          end.line,
-          end.character + trailingDelimiters[0].length
-        )
-      : null;
-
-  const isInDelimitedList =
-    (leadingDelimiterRange != null || trailingDelimiterRange != null) &&
-    (leadingDelimiterRange != null || start.character === 0) &&
-    (trailingDelimiterRange != null || end.isEqual(endLine.range.end));
-
-  return {
-    isInDelimitedList,
-    containingListDelimiter: " ",
-    leadingDelimiterRange: isInDelimitedList ? leadingDelimiterRange : null,
-    trailingDelimiterRange: isInDelimitedList ? trailingDelimiterRange : null,
-    outerSelection: selectionContext.outerSelection,
-  };
-}
-
-// TODO Clean this up once we have rich targets and better polymorphic
-// selection contexts that indicate their type
-function isSelectionContextEmpty(selectionContext: SelectionContext) {
-  return (
-    selectionContext.isInDelimitedList == null &&
-    selectionContext.containingListDelimiter == null &&
-    selectionContext.leadingDelimiterRange == null &&
-    selectionContext.trailingDelimiterRange == null
-  );
-}
-
-function getLineSelectionContext(
-  selection: SelectionWithEditor
-): SelectionContext {
-  const { document } = selection.editor;
-  const { start, end } = selection.selection;
-  const outerSelection = getOuterSelection(selection, start, end);
-
-  const leadingDelimiterRange =
-    start.line > 0
-      ? new Range(
-          document.lineAt(start.line - 1).range.end,
-          outerSelection.start
-        )
-      : null;
-  const trailingDelimiterRange =
-    end.line + 1 < document.lineCount
-      ? new Range(outerSelection.end, new Position(end.line + 1, 0))
-      : null;
-  const isInDelimitedList =
-    leadingDelimiterRange != null || trailingDelimiterRange != null;
-
-  return {
-    isInDelimitedList,
-    containingListDelimiter: "\n",
-    leadingDelimiterRange,
-    trailingDelimiterRange,
-    outerSelection,
-  };
-}
-
-function getParagraphSelectionContext(
-  selection: SelectionWithEditor
-): SelectionContext {
-  const { document } = selection.editor;
-  const { start, end } = selection.selection;
-  const outerSelection = getOuterSelection(selection, start, end);
-  const leadingLine = getPreviousNonEmptyLine(document, start.line);
-  const trailingLine = getNextNonEmptyLine(document, end.line);
-
-  const leadingDelimiterStart =
-    leadingLine != null
-      ? leadingLine.range.end
-      : start.line > 0
-      ? new Position(0, 0)
-      : null;
-  const trailingDelimiterEnd =
-    trailingLine != null
-      ? trailingLine.range.start
-      : end.line < document.lineCount - 1
-      ? document.lineAt(document.lineCount - 1).range.end
-      : null;
-  const leadingDelimiterRange =
-    leadingDelimiterStart != null
-      ? new Range(leadingDelimiterStart, outerSelection.start)
-      : null;
-  const trailingDelimiterRange =
-    trailingDelimiterEnd != null
-      ? new Range(outerSelection.end, trailingDelimiterEnd)
-      : null;
-  const isInDelimitedList =
-    leadingDelimiterRange != null || trailingDelimiterRange != null;
-
-  return {
-    isInDelimitedList,
-    containingListDelimiter: "\n\n",
-    leadingDelimiterRange,
-    trailingDelimiterRange,
-    outerSelection,
-  };
-}
-
-function getOuterSelection(
-  selection: SelectionWithEditor,
-  start: Position,
-  end: Position
-) {
-  // Outer selection contains the entire lines
-  return selectionFromPositions(
-    selection.selection,
-    new Position(start.line, 0),
-    selection.editor.document.lineAt(end).range.end
-  );
-}
-
-function getPreviousNonEmptyLine(
-  document: TextDocument,
-  startLineNumber: number
-) {
-  let line = document.lineAt(startLineNumber);
-  while (line.lineNumber > 0) {
-    const previousLine = document.lineAt(line.lineNumber - 1);
-    if (!previousLine.isEmptyOrWhitespace) {
-      return previousLine;
-    }
-    line = previousLine;
-  }
-  return null;
-}
-
-function getNextNonEmptyLine(document: TextDocument, startLineNumber: number) {
-  let line = document.lineAt(startLineNumber);
-  while (line.lineNumber + 1 < document.lineCount) {
-    const nextLine = document.lineAt(line.lineNumber + 1);
-    if (!nextLine.isEmptyOrWhitespace) {
-      return nextLine;
-    }
-    line = nextLine;
-  }
-  return null;
 }
